[THORN-2563] enforce dependency convergence
This commit adds Maven Enforcer plugin for the main codebase
(not for the test suite) and adds a bunch of dependency exclusions
so that `mvn validate` succeeds.

There should be no behavioral changes.

I'm aware of one possibly breaking change: the `jaxrs` fraction
used to provide the JAXB API, but no longer does. That was
arguably a bug and users should use `jaxrs-jaxb` anyway.