THORN-2585: WildFly OpenSSL doesn't work
Motivation
----------
When WildFly OpenSSL is enabled, Thorntail applications fail to boot.
This is because we don't include the `libwfssl` native libraries.

Modifications
-------------
Instead of `org.wildfly.openssl:wildfly-openssl-java`, which is
just plain Java code (and the native libraries must be provided
separately), we now use `org.wildfly.openssl:wildfly-openssl`,
which is the same Java code + the compiled native libraries.

In addition to that, this commit also fixes how the Arquillian
adapter stops the running application. It first tries to delete
the process file, but that could never possibly have worked,
because the process file is created using `TempFileManager`,
which used to always delete the temporary file before handing
back its path. Almost all users of `TempFileManager` are ready
to deal with newly created temp files existing, so this commit
changes the code to _not_ delete the file. The one caller
expecting the temp file to not exist was updated to delete it
explicitly.

Result
------
WildFly OpenSSL works as intended.
Stopping the running application by deleting the process file
also works as intended.